\name{DSQ}
\alias{DSQ}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Scoring of the 26-item Dietary Screener Questionnaire (DSQ)
}
\description{

}
\usage{
Score_DSQ(data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
}
\details{
}
\value{
}
\references{
}
\author{
Wangcheng Yan, Wenjun Zhou
}
\note{
}
\section{Warning }{
}
\seealso{
}
\examples{
}
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
