% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.dsq.self.administer.paper.R
\name{DSQ_scores}
\alias{DSQ_scores}
\title{Scoring DSQ}
\usage{
DSQ_scores(data, method = "current", keep = TRUE)
}
\arguments{
\item{data}{The input data frame.}

\item{method}{The version of scoring algorithm. Choose from "current" or "earlier". The default is "current".}

\item{keep}{Whether to keep the variables in the input data frame. The default is TRUE.}
}
\value{
The intake values for calcium, etc.
}
\description{
This function calculates.
}
\examples{
data_scored <- DSQ_scores(data,method="c");
}
