% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAT.R
\name{FAT_scores}
\alias{FAT_scores}
\title{Scoring Fat}
\usage{
FAT_scores(data)
}
\arguments{
\item{data}{The input data frame.}
}
\value{
The intake values for fat, etc.
}
\description{
This function calculates.
}
\examples{
df_out <- FAT_scores(df_in);
}
