% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FV.R
\name{FV_scores}
\alias{FV_scores}
\title{Scoring FV}
\usage{
FV_scores(data)
}
\arguments{
\item{data}{The input data frame.}
}
\value{
The intake values for fruits and vegetables, etc.
}
\description{
This function calculates.
}
\examples{
df_out <- FV_scores(df_in);
}
