% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEI2015perpersonNDSR2013orsubsequent.R
\name{HEI_scores}
\alias{HEI_scores}
\title{Scoring HEI from NDSR}
\usage{
HEI_scores(data)
}
\arguments{
\item{data}{The input data frame.}
}
\value{
The HEI scores, etc.
}
\description{
This function calculates.
}
\examples{
df_out <- HEI_scores(df_in);
}
