\name{read_TSV}
\alias{read_TSV}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Read TSV File
}
\description{
Read data (Note: colClasses = "character" option ensures that the 
data are all character type, and missing values are all coded as 
"" - empty strings)
}
\usage{
read_TSV(filepathname)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{filepathname}{file path}
}
\format{
}
\details{
}
\value{
}
\references{
}
\author{
Wangcheng Yan, Wenjun Zhou
}
\note{
}
\section{Warning }{
}
\seealso{
}
\examples{
}
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
