% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sHEI.R
\name{sHEI_scores}
\alias{sHEI_scores}
\title{Comput sHEI scores}
\usage{
sHEI_scores(df, miss_val = 0)
}
\arguments{
\item{df}{The data frame with sHEI questionnaire responses}

\item{miss_val}{The value to use for missing data}
}
\value{
A data frame with computed columns in addition
}
\description{
Compute the estimated HEI scores using sHEI questionnaire responses
}
\examples{
df_scored <- sHEI_scores(df);
df_scored <- sHEI_scores(df,miss_val=0);
}
