\name{Stress14}
\alias{Stress14}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Scoring of the ``Stress14'' survey
}
\description{
The function return the stress score given the data
}
\usage{
Stress14(data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
    Your input data containing the responses of the stress survey
  }
}
\details{
  The variable names of the stress survey should be as follows, the ordering does not matter.
  \begin{itemize}
    \item a
  \end{itemize}
}
\value{
  ~Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Wangcheng Yan, Wenjun Zhou
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
