\name{zBMI}
\alias{zBMI}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Scoring of zBMI and zBMI category
}
\description{

}
\usage{

}
%- maybe also 'usage' for other objects documented here.
\arguments{
}
\format{
}
\details{
}
\value{
}
\references{
}
\author{
Wangcheng Yan, Wenjun Zhou
}
\note{
}
\section{Warning }{
}
\seealso{
}
\examples{
}
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
